refactor: rename nut.config.js to pages.config.js, change commands to nut pages dev and nut pages build
BREAKING CHANGE: rename nut.config.js to pages.config.js, change commands to nut pages dev and nut pages build
