+++ examples/child1/pages.config.js
