+++ examples/child2/pages.config.js
