+++ examples/host/pages.config.js
