+++ examples/now2/pages.config.js
