+++ examples/simple/package.json
@@ -4,8 +4,8 @@
   "version": "1.1.3-alpha.0",
   "description": "",
   "scripts": {
-    "dev": "node ../../packages/cli/bin/index.js",
-    "build": "node ../../packages/cli/bin/index.js --prod"
+    "dev": "node ../../packages/cli/bin/index.js pages dev",
+    "build": "node ../../packages/cli/bin/index.js pages build"
   },
   "keywords": [],
   "author": "",
