+++ examples/simple/pages.config.js
