+++ package.json
@@ -15,7 +15,7 @@
     "dev": "npm run dev --prefix ./examples/simple",
     "build": "npm run build --prefix ./examples/simple",
     "dev:now2": "npm run dev --prefix ./examples/now2",
-    "c": "node ./packages/cli/bin/index.js create demo",
+    "c": "node ./packages/cli/bin/index.js pages create demo",
     "build:compose": "npm run build --prefix ./examples/host"
   },
   "husky": {
@@ -33,7 +33,7 @@
   "workspaces": {
     "packages": [
       "packages/*",
-      "plugins/*",
+      "plugins/pages/*",
       "examples/*",
       "website"
     ]
