--- packages/driver-pages/lib/commands/index.js
@@ -1,5 +1,0 @@
-module.exports = {
-  development: require( './dev' ),
-  production: require( './prod' ),
-  create: require( './create' ),
-}
