+++ packages/driver-pages/lib/index.js
@@ -4,6 +4,5 @@ const pkg = require( '../package.json' )
 exports.name = pkg.name
 exports.version = pkg.version
 exports.apply = function ( { register } ) {
-  register( commands )
-  register( commands )
+  register( 'pages', commands )
 }
