+++ website/package.json
@@ -3,8 +3,8 @@
   "version": "0.1.1-alpha.1",
   "private": true,
   "scripts": {
-    "dev": "node ../packages/cli/bin/index.js",
-    "build": "node ../packages/cli/bin/index.js --prod"
+    "dev": "node ../packages/cli/bin/index.js pages dev",
+    "build": "node ../packages/cli/bin/index.js pages build"
   },
   "dependencies": {
     "@zeit-ui/vue": "1.3.1"
